Refine A* tracing routine.
Former-commit-id: 89a8b726b7253cb0eaee8d9397309e089f02725d

+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/slice_viewer/Subvolume.java
@@ -270,7 +270,7 @@ public class Subvolume {
         return extent;
     }
 
-    public long getIntensityGlobal(ZoomedVoxelIndex v1, int channelIndex) 
+    public int getIntensityGlobal(ZoomedVoxelIndex v1, int channelIndex) 
     {
         int c = channelIndex;
         int x = v1.getX() - origin.getX();
